Added new stage: L4_graph_runner
This stage will add tests from L4 over time to maintain correct working of the tests.
Currently there are only a handful of tests, but more will be added as support expands. Eventually we will either move this stage out of gitlab, or add more gitlab runners.

Note also that we intentionally did not add the stage for Hopper as there were some issues with accuracy. Once resolved, Hopper will be added to the list.