Option to set max KV sequence length explicitly
Three changes:

* Add back the cpp sample for paged caches
* Rename the paged attention related options
* Add the option set max KV sequence length explicitly

Reason for the latter:

**TL;DR: The issue is that we are leaving a trap behind for developers of any feature that requires input tensors with s_kv as one of the dimensions.** 

We have all kind of ways to infer s_kv: from the bias, from the RNG/dropout, and if none of those are present, we fall back to an overestimate of S_KV. I am anticipating that as we add new features to sdpa, there will be more ways to infer s_kv. Not adding this to the list of ways to infer s_kv will result in conflicts with our fallback estimate.Sure enough, with Flex Attention coming from behind the corner, this confirmed my point: Flex attention could be using tensors that use s_kv as one of the dimensions. We don't want future devs to keep updating how we infer s_kv in the case of paged attention. And neither do we want our "overestimated s_kv" conflicts with the actual s_kv from new features.

Commit history:

* Addressing windows build issue

* Addressed suggestion Ane

* Added a warning message

* Added back paged caches CPP sample that was removed

Commit 152a0a3b044ed7bb49bd5c87f3bc2009d91009e2 accidentally removed
samples/cpp/sdpa/fp16_fwd_with_paged_caches.cpp from being compiled.

This commit adds it back the the CMakeLists and adds the necessary
adjustments.

* Renamed paged attention vars and fixed small issues

* Typo

* Option to set max KV sequence length explicitly