Improving tests blocking in ‘test_mhas_v2.py’; fixing bug https://nvbugs/5390904
* Improving tests blocking in ‘test_mhas_v2.py’.  The syntax of the .block file does not change.
However, multiple definitions for the same test name are allowed.  The fetch_blocked_tests()
function reads 'test_mhas_v2.block' and checks if any definition applies to the current GPU
architecture and cuDNN version.  If so, the test name is added to the list of blocked tests -
a simple list of strings.

Fixing bug https://nvbugs/5390904: get_device_properties() missing 1 required positional
argument: 'device'.  The 'device' argument should be optional in PyTorch but apparently,
this is not always the case.